requirements-dev.txt
-r requirements.txt
pytest >= 4.1.1
pytest-datadir>=1.3
pytest-factoryboy>=2.0.3
factory-boy>=2.11
vcrpy >= 2.0.1
mypy>=0.720
mypy-extensions>=0.4.1
sqlalchemy-stubs>=0.2
